refactor: extract Powerup class to its own module
- Moved Powerup class from sprites/sprites.py to sprites/powerup.py
- Added import for Powerup in sprites/__init__.py
- Verified powerup spawning and movement
