+++ sprites/__init__.py
@@ -2,3 +2,4 @@ from .sprites import *
 from .starfield import Starfield
 from .meteoroid import Meteoroid
 from .bullet import Bullet
+from .powerup import Powerup
